requirements.txt
python-vlc
mpris-server
